Make fnCollapseSA convert TimeTree -> TimeTree
And BranchLengthTree -> BranchLengthTree.
